Create global ConfigStore class
Used to read in configuration from a file and allow global access

The global object is readonly after initial creation to prevent possible configuration errors at runtime.  As a result, changing filebased configuration will require restarting the system.

The type variable is to denote whether the server is running as a provisioning or handling server.