Add Timer Retransmission
Timer now retransmits data at a fixed interval

When UDP::Handle is passed in as a shared pointer to the lambda expression, you need to pass it in as a copy rather than a reference.  Otherwise, the data is destroyed (probably because its a shared_ptr).

Change globalConfig to a project wide global variable and extern it in config.hpp.  It is still readonly.