Add batch save functionality
Batch saves are done via calculating the number of frequencies that need to be saved and then appending that many additions to the prepared statements.

Fix a use after free bug affecting linux where that char* passed to udp->send would sometimes be freed before the data was actually sent.  The data is now passed as a unique_ptr to avoid leaking memoery

Added rudimentary build instructions text file to track how to build the system on Linux.