Fix packet timestamp overlap
Add several checks to avoid overlapping timestamps on incoming messages.  This could result in several conflicting values for the same frequency being saved to the database.

The process flow is:
If the packet has arrived within the designated duration specified, check if its frequency is already in the buffer.  If it is, double check to confirm whether or not it can be placed in a future buffer.